builtin.inc: fix build with -Woverlength-strings
C99 only allows string literals with length up to 4095 characters.

This patch introduces a gen_jq_builtins generator program that generates
the jq_builtins array in builtin.c as an array literal instead of a
string literal to make compilation not fail when building with
-Woverlength-strings.

Add -Woverlength-strings to CFLAGS to verify that the fix works.

Also introduce HOSTCC autoconf variable to get a C compiler to build
gen_jq_builtin on the host when cross-compiling.

Fixes #1481