Add a notion of maximum depth to the parser state and `jq_state`.
It's not sufficient to just add the state to the parser, since we
might call `jv_load_file` from a variety of other functions.

A few helper functions defer to `DEFAULT_MAX_PARSING_DEPTH`, as they
are given neither a `jq_state` nor `jv_parser`.

Added tests in `tests/shtest` to confirm that the maximum depth limit
is appropriately enforced.