Improve error messages to show problematic position in the filter
This commit improves the error messages of parser and compiler to make
it more easy to understand where the filter problem is. Although the
error messages include the problematic line in the filter, it is
sometimes difficult to understand where in the line is the issue.
By decorating problematic token with underline, it becomes very easy.